暫存

+++ src/VideoRent/MovieTypes/MovieTypes.java
@@ -3,6 +3,6 @@ package VideoRent.MovieTypes;
 public interface MovieTypes {
 	
 	public float CustomerPoint();
-	public int RentPrice(int days);
+	public float RentPrice(int days);
 
 }
